feat: split kubeseal logic into multiple classes
